feat(src): update all dependencies
Update to history 3.x and all dev dependencies for testing
