+++ package.json
@@ -18,34 +18,36 @@
   },
   "homepage": "https://github.com/cyclejs/history#readme",
   "dependencies": {
-    "history": "^2.1.0"
+    "history": "^3.0.0"
   },
   "devDependencies": {
     "@cycle/base": "^3.0.0",
-    "@cycle/dom": "^10.0.0-rc8",
+    "@cycle/dom": "^10.0.0-rc28",
+    "@cycle/most-adapter": "^2.0.1",
+    "@cycle/most-run": "^2.0.0",
     "@cycle/rx-adapter": "^2.0.0",
     "@cycle/rxjs-adapter": "^2.0.0",
     "@cycle/xstream-adapter": "^2.0.0",
     "@cycle/xstream-run": "^2.0.0",
-    "assert": "^1.3.0",
-    "babel-preset-es2015": "^6.6.0",
-    "babel-register": "^6.7.2",
-    "babelify": "^7.2.0",
-    "browserify": "^13.0.0",
+    "assert": "^1.4.1",
+    "babel-preset-es2015": "^6.9.0",
+    "babel-register": "^6.9.0",
+    "babelify": "^7.3.0",
+    "browserify": "^13.0.1",
     "cli-release": "^1.0.4",
     "ghooks": "^1.2.1",
     "mkdirp": "^0.5.1",
-    "mocha": "^2.4.5",
+    "mocha": "^2.5.3",
+    "most": "^0.19.7",
     "rx": "^4.1.0",
-    "rxjs": "^5.0.0-beta.7",
-    "testem": "^1.6.0",
-    "ts-node": "^0.7.2",
-    "tslint": "^3.8.0",
+    "rxjs": "^5.0.0-beta.8",
+    "testem": "^1.8.1",
+    "tslint": "^3.10.2",
     "typescript": "^1.8.10",
     "typings": "^0.8.1",
     "uglify-js": "^2.6.2",
     "validate-commit-message": "^3.0.1",
-    "xstream": "^4.0.0"
+    "xstream": "^4.0.1"
   },
   "config": {
     "ghooks": {
