+++ src/interfaces.ts
@@ -27,7 +27,7 @@ export interface History {
   replace(location: Location| Pathname): void;
   createHref(href: Pathname): string;
   createLocation(location: Location | Pathname): Location;
-
+  getCurrentLocation(): Location;
   addCompleteCallback?: (fn: () => void) => void;
   complete?: () => void;
 };
